Puliendo la UI y los BO, agregando validaciones

+++ src/ar/com/rorra/dao/DoctorDAO.java
@@ -78,7 +78,7 @@ public class DoctorDAO extends BaseDAO<Doctor> implements IDoctorDAO {
    * @return
    * @throws DBException
    */
-  public Doctor getByLegajo(long legajo) throws DBException {
+  public Doctor getByLegajo(int legajo) throws DBException {
     return getByField("legajo", String.valueOf(legajo));
   }
